one more

+++ Source/Misc/Polytempo_TempoMeasurement.h
@@ -15,7 +15,7 @@ public:
 
         // round to max. 3 decimal places
         float value = absoluteTempo / Rational(tokens[0]).toFloat() * tokens[1].getFloatValue();
-        value = int(value * 1000.0f + 0.5f);
+        value = (float)(int(value * 1000.0f + 0.5f));
         return value / 1000.0f;
     }
